+++ bin/main.ml
@@ -1480,6 +1480,43 @@ let printenv =
   in
   (term, Term.info "printenv" ~doc ~man )
 
+let fmt =
+  let doc = "Format dune files" in
+  let man =
+    [ `S "DESCRIPTION"
+    ; `P {|$(b,dune unstable-fmt) reads a dune file and outputs a formatted
+           version. This feature is unstable, and its interface or behaviour
+           might change.
+         |}
+    ] in
+  let term =
+    let%map path_opt =
+      let docv = "FILE" in
+      let doc = "Path to the dune file to parse." in
+      Arg.(value & pos 0 (some path) None & info [] ~docv ~doc)
+    and inplace =
+      let doc = "Modify the file in place" in
+      Arg.(value & flag & info ["inplace"] ~doc)
+    in
+    if true then
+      let (input, output) =
+        match path_opt, inplace with
+        | None, false ->
+          (None, None)
+        | Some path, true ->
+          let path = Arg.Path.path path in
+          (Some path, Some path)
+        | Some path, false ->
+          (Some (Arg.Path.path path), None)
+        | None, true ->
+          die "--inplace requires a file name"
+      in
+      Dune_fmt.format_file ~input ~output
+    else
+      die "This command is unstable. Please pass --unstable to use it nonetheless."
+  in
+  (term, Term.info "unstable-fmt" ~doc ~man )
+
 module Help = struct
   let config =
     ("dune-config", 5, "", "Dune", "Dune manual"),
@@ -1600,6 +1637,7 @@ let all =
   ; promote
   ; printenv
   ; Help.help
+  ; fmt
   ]
 
 let default =